integrada la nueva versi?n de qtwebapp

+++ YACReaderLibrary/server/lib/bfLogging/logmessage.cpp
@@ -49,7 +49,7 @@ QString LogMessage::toString(const QString& msgFormat, const QString& timestampF
     }
 
     QString threadId;
-    threadId.setNum((qint64)QThread::currentThreadId());
+    threadId.setNum((unsigned int)QThread::currentThreadId());
     decorated.replace("{thread}",threadId);
 
     // Fill in variables
